+++ lib/gamefic/command.rb
@@ -35,6 +35,14 @@ module Gamefic
       @precision = precision
     end
 
+    def substantiality
+      @substantiality ||= arguments.that_are(Entity).length + (verb ? 1 : 0)
+    end
+
+    def inspect
+      "#<#{self.class} #{([verb] + arguments).map(&:inspect).join(', ')}>"
+    end
+
     class << self
       # Compose a command from input.
       #
@@ -53,11 +61,12 @@ module Gamefic
       # @param expression [Expression]
       # @return [Array<Command>]
       def expression_to_commands actor, expression
+        Gamefic.logger.info "Evaluating #{expression.inspect}"
         actor.epic
              .responses_for(expression.verb)
              .map { |response| response.to_command(actor, expression) }
              .compact
-             .sort_by.with_index { |result, idx| [-result.precision, -result.strictness, idx] }
+             .sort_by.with_index { |result, idx| [-result.substantiality, -result.strictness, -result.precision, idx] }
       end
     end
   end