+++ lib/gamefic/expression.rb
@@ -17,6 +17,9 @@ module Gamefic
       @tokens = tokens
     end
 
+    def inspect
+      "#<#{self.class} #{([verb] + tokens).map(&:inspect).join(', ')}>"
+    end
     # Compare two syntaxes for the purpose of ordering them by relevance while
     # dispatching.
     #