+++ lib/gamefic/proxy.rb
@@ -1,7 +1,16 @@
 # frozen_string_literal: true
 
 module Gamefic
+  # @todo Turn this into a module after the old proxies are completely deprecated
+  #
   class Proxy
+    require 'gamefic/proxy/base'
+    require 'gamefic/proxy/config'
+    require 'gamefic/proxy/pick'
+    require 'gamefic/proxy/plot_pick'
+
+    TYPES = %i[attr ivar pick pick! plot_pick plot_pick! config].freeze
+
     # @return [Symbol]
     attr_reader :type
 
@@ -9,16 +18,24 @@ module Gamefic
     attr_reader :key
 
     # @param type [Symbol]
-    # @param key [Symbol, String]
+    # @param key [Symbol, String, Array]
     def initialize type, key
+      Gamefic.logger.debug "Using deprecated #{type} proxy"
       @type = type
-      @key = key
       validate_type
+      @key = type == :config ? [key].compact : key
     end
 
     def fetch narrative
       send(type, narrative) ||
         raise(ArgumentError, "Unable to fetch entity from proxy agent symbol `#{key}`")
+    end
+
+    def [](key)
+      raise ArgumentError, 'Invalid []' unless type == :config
+
+      @key.push key
+      self
     end
 
     private
@@ -34,13 +51,29 @@ module Gamefic
     end
 
     def pick narrative
-      narrative.pick! key
+      narrative.pick *key
+    end
+
+    def pick! narrative
+      narrative.pick! *key
+    end
+
+    def plot_pick narrative
+      narrative.plot.pick *key
+    end
+
+    def plot_pick! narrative
+      narrative.plot.pick! *key
+    end
+
+    def config narrative
+      key.inject(narrative.config) { |hash, key| hash[key] }
     end
 
     def validate_type
-      return if [:attr, :ivar, :pick].include?(type)
+      return if TYPES.include?(type)
 
-      raise ArgumentError, "Invalid proxy type `#{type}` (must be :attr, :ivar, or :pick)"
+      raise ArgumentError, "Invalid proxy type `#{type}` (must be #{TYPES.join_or})"
     end
   end
 end