+++ lib/gamefic/query/scoped.rb
@@ -10,14 +10,13 @@ module Gamefic
       attr_reader :scope
 
       # @param scope [Class<Gamefic::Scope::Base>]
-      def initialize scope, *arguments, ambiguous: false
-        super(*arguments, ambiguous: ambiguous)
+      def initialize scope, *arguments, ambiguous: false, name: nil
+        super(*arguments, ambiguous: ambiguous, name: name)
         @scope = scope
       end
 
-      def select(subject)
+      def span(subject)
         @scope.matches(subject)
-              .that_are(*unproxied_arguments)
       end
 
       def precision