+++ lib/gamefic/scanner/nesting.rb
@@ -12,7 +12,7 @@ module Gamefic
       end
 
       def scan
-        return Result.unmatched(selection, token, self.class) unless token =~ NEST_REGEXP
+        return unmatched_result unless token =~ NEST_REGEXP
 
         denest selection, token
       end
@@ -27,7 +27,7 @@ module Gamefic
           current = parts.pop
           last_result = subprocessor.scan(near, current)
           last_result = subprocessor.scan(far, current) if last_result.matched.empty? && near != far
-          return Result.unmatched(selection, token, self.class) if last_result.matched.empty? || last_result.matched.length > 1
+          return unmatched_result if last_result.matched.empty? || last_result.matched.length > 1
 
           near = last_result.matched.first.children & objects
           far = last_result.matched.first.flatten & objects