+++ lib/gamefic/scriptable/queries.rb
@@ -12,7 +12,18 @@ module Gamefic
       # @param args [Array<Object>] Query arguments
       # @return [Query::General]
       def anywhere *args, ambiguous: false
-        Query::General.new -> { entities }, *args, ambiguous: ambiguous
+        Query::General.new -> { entities }, *args, ambiguous: ambiguous, name: 'anywhere'
+      end
+
+      # Define a query that searches for abstract entities.
+      #
+      # An abstract entity is a pseudo-entity that is describable but does
+      # not have a parent or children.
+      #
+      # @param args [Array<Object>] Query arguments
+      # @return [Query::Abstract]
+      def abstract *args, ambiguous: false
+        Query::Abstract.new -> { entities }, *args, ambiguous: ambiguous
       end
 
       # Define a query that searches an actor's family of entities. The
@@ -22,7 +33,7 @@ module Gamefic
       # @param args [Array<Object>] Query arguments
       # @return [Query::Scoped]
       def available *args, ambiguous: false
-        Query::Scoped.new Scope::Family, *args, ambiguous: ambiguous
+        Query::Scoped.new Scope::Family, *args, ambiguous: ambiguous, name: 'available'
       end
       alias family available
 
@@ -31,7 +42,7 @@ module Gamefic
       # @param args [Array<Object>] Query arguments
       # @return [Query::Scoped]
       def parent *args, ambiguous: false
-        Query::Scoped.new Scope::Parent, *args, ambiguous: ambiguous
+        Query::Scoped.new Scope::Parent, *args, ambiguous: ambiguous, name: 'parent'
       end
 
       # Define a query that searches an actor's children.
@@ -39,7 +50,7 @@ module Gamefic
       # @param args [Array<Object>] Query arguments
       # @return [Query::Scoped]
       def children *args, ambiguous: false
-        Query::Scoped.new Scope::Children, *args, ambiguous: ambiguous
+        Query::Scoped.new Scope::Children, *args, ambiguous: ambiguous, name: 'children'
       end
 
       # Define a query that searches an actor's descendants.
@@ -55,7 +66,7 @@ module Gamefic
       # @param args [Array<Object>] Query arguments
       # @return [Query::Scoped]
       def siblings *args, ambiguous: false
-        Query::Scoped.new Scope::Siblings, *args, ambiguous: ambiguous
+        Query::Scoped.new Scope::Siblings, *args, ambiguous: ambiguous, name: 'siblings'
       end
 
       # Define a query that returns the actor itself.
@@ -63,7 +74,7 @@ module Gamefic
       # @param args [Array<Object>] Query arguments
       # @return [Query::Scoped]
       def myself *args, ambiguous: false
-        Query::Scoped.new Scope::Myself, *args, ambiguous: ambiguous
+        Query::Scoped.new Scope::Myself, *args, ambiguous: ambiguous, name: 'myself'
       end
 
       # Define a query that performs a plaintext search. It can take a String
@@ -74,7 +85,7 @@ module Gamefic
       # @param arg [String, Regexp] The string or regular expression to match
       # @return [Query::Text]
       def plaintext arg = /.*/
-        Query::Text.new arg
+        Query::Text.new arg, name: 'plaintext'
       end
     end
   end