+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.6.0 - October 6, 2024
+- Normalized arguments accept strings
+- Smarter picks and proxies
+- Commands prefer strictness over precision
+- Queries scan for ambiguity before filtering through arguments
+- Abstract queries
+- Command logging
+
 ## 3.5.0 - October 5, 2024
 - Configurable scanners
 - Refactored scanners and queries